Fix signed-int cycle-counter overflow that froze giant traces
The interconnect edge counters (DirectInterconnect/Bus/Mesh) were signed
int. They gate message movement on `edges % 2 == 1`; at 2^31 edges
(~1.07B cycles) the counter wrapped negative, making the modulo never 1,
so the interconnect stopped moving messages and the sim deadlocked
periodically (~1.07B and ~3.22B cycles). Widen the cycle/edge counters to
uint64_t so the modulo stays 0/1 and wraps cleanly.

Also widen CacheDataHandler m_cycle/m_ready_cycle and CPU
m_last_received_msg_cycle to uint64_t: the SPLASH giants exceed 2^32
cycles, where the old uint32_t stamps wrapped.

With this, raytrace/radiosity/radix run to completion across all
protocols (snoop + directory x MSI/MESI/MOESI).

+++ header/CPU.h
@@ -43,9 +43,9 @@ namespace octopus
         int m_id;
         uint64_t m_clk_cycle;
         uint32_t m_number_of_OoO_requests;
-        uint32_t m_last_received_msg_cycle;
+        uint64_t m_last_received_msg_cycle; // 64-bit: wraps past 2^32 on giant traces
         int32_t m_sent_requests;
-        bool m_simulation_done; 
+        bool m_simulation_done;
 
         TraceSample* m_sample_in_progess;
         std::ifstream  m_workload_file;
